fix #12: Compatible with Openfire 5.0.0
This replaces usage of Openfire API that was dropped in 5.0.0 with API that was introduced in 4.4.0. Also, a third-party library that no longer ships with Openfire is now added as a direct dependency of this plugin.

As a result, the plugin should now be compatible with Openfire 4.4.0 and later (including 5.0.0).